+++ packages/reactabular-table/CHANGELOG.md
@@ -1,3 +1,8 @@
+1.2.0 / 2016-08-05
+==================
+
+  * Bug fix - Pass unresolved values to `Table.Body` transforms instead of resolved ones.
+
 1.1.1 / 2016-08-04
 ==================